+++ engine/src/test/java/net/jqwik/api/edgeCases/ArbitraryEdgeCasesTests.java
@@ -30,10 +30,10 @@ class ArbitraryEdgeCasesTests {
 		Arbitrary<String> arbitrary = Arbitraries.integers().between(-10, 10).map(i -> Integer.toString(i));
 		EdgeCases<String> edgeCases = arbitrary.edgeCases();
 		assertThat(values(edgeCases)).containsExactlyInAnyOrder(
-			"-10", "-2", "-1", "0", "1", "2", "10"
+			"-10", "-9", "-2", "-1", "0", "1", "2", "9", "10"
 		);
 		// make sure edge cases can be repeatedly generated
-		assertThat(values(edgeCases)).hasSize(7);
+		assertThat(values(edgeCases)).hasSize(9);
 	}
 
 	@Example
@@ -68,13 +68,13 @@ class ArbitraryEdgeCasesTests {
 
 	@Example
 	void injectNull() {
-		Arbitrary<Integer> arbitrary = Arbitraries.integers().between(-10, 10).injectNull(0.1);
+		Arbitrary<Integer> arbitrary = Arbitraries.of(-10, 10).injectNull(0.1);
 		EdgeCases<Integer> edgeCases = arbitrary.edgeCases();
 		assertThat(values(edgeCases)).containsExactlyInAnyOrder(
-			null, -10, -2, -1, 0, 1, 2, 10
+			null, -10, 10
 		);
 		// make sure edge cases can be repeatedly generated
-		assertThat(values(edgeCases)).hasSize(8);
+		assertThat(values(edgeCases)).hasSize(3);
 	}
 
 	@Example
@@ -82,10 +82,10 @@ class ArbitraryEdgeCasesTests {
 		Arbitrary<Integer> arbitrary = Arbitraries.integers().between(-10, 10).fixGenSize(100);
 		EdgeCases<Integer> edgeCases = arbitrary.edgeCases();
 		assertThat(values(edgeCases)).containsExactlyInAnyOrder(
-			-10, -2, -1, 0, 1, 2, 10
+			-10, -9, -2, -1, 0, 1, 2, 9, 10
 		);
 		// make sure edge cases can be repeatedly generated
-		assertThat(values(edgeCases)).hasSize(7);
+		assertThat(values(edgeCases)).hasSize(9);
 	}
 
 	@Example
@@ -93,36 +93,34 @@ class ArbitraryEdgeCasesTests {
 		Arbitrary<Integer> arbitrary = Arbitraries.integers().between(-10, 10).unique();
 		EdgeCases<Integer> edgeCases = arbitrary.edgeCases();
 		assertThat(values(edgeCases)).containsExactlyInAnyOrder(
-			-10, -2, -1, 0, 1, 2, 10
+			-10, -9, -2, -1, 0, 1, 2, 9, 10
 		);
 		// make sure edge cases can be repeatedly generated
-		assertThat(values(edgeCases)).hasSize(7);
+		assertThat(values(edgeCases)).hasSize(9);
 	}
 
 	@Example
 	void flatMapping() {
-		Arbitrary<Integer> arbitrary = Arbitraries.integers()
-												  .between(5, 10)
+		Arbitrary<Integer> arbitrary = Arbitraries.of(5, 10)
 												  .flatMap(i -> Arbitraries.integers().between(i, i * 10));
 
 		EdgeCases<Integer> edgeCases = arbitrary.edgeCases();
 		assertThat(values(edgeCases)).containsExactlyInAnyOrder(
-			5, 50, 10, 100
+			5, 6, 49, 50, 10, 11, 99, 100
 		);
 		// make sure edge cases can be repeatedly generated
-		assertThat(values(edgeCases)).hasSize(4);
+		assertThat(values(edgeCases)).hasSize(8);
 	}
 
 	@Example
 	void optionals() {
-		Arbitrary<Optional<Integer>> arbitrary = Arbitraries.integers().between(-10, 10).optional();
+		Arbitrary<Optional<Integer>> arbitrary = Arbitraries.of(-10, 10).optional();
 		EdgeCases<Optional<Integer>> edgeCases = arbitrary.edgeCases();
 		assertThat(values(edgeCases)).containsExactlyInAnyOrder(
-			Optional.empty(), Optional.of(-10), Optional.of(-2), Optional.of(-1), Optional.of(0), Optional.of(1), Optional.of(2), Optional
-																																	  .of(10)
+			Optional.empty(), Optional.of(-10), Optional.of(10)
 		);
 		// make sure edge cases can be repeatedly generated
-		assertThat(values(edgeCases)).hasSize(8);
+		assertThat(values(edgeCases)).hasSize(3);
 	}
 
 	@Group
@@ -130,20 +128,15 @@ class ArbitraryEdgeCasesTests {
 
 		@Example
 		void listEdgeCases() {
-			IntegerArbitrary ints = Arbitraries.integers().between(-10, 10);
+			Arbitrary<Integer> ints = Arbitraries.of(-10, 10);
 			Arbitrary<List<Integer>> arbitrary = ints.list();
 			assertThat(values(arbitrary.edgeCases())).containsExactlyInAnyOrder(
 				Collections.emptyList(),
 				Collections.singletonList(-10),
-				Collections.singletonList(-2),
-				Collections.singletonList(-1),
-				Collections.singletonList(0),
-				Collections.singletonList(1),
-				Collections.singletonList(2),
 				Collections.singletonList(10)
 			);
 			// make sure edge cases can be repeatedly generated
-			assertThat(values(arbitrary.edgeCases())).hasSize(8);
+			assertThat(values(arbitrary.edgeCases())).hasSize(3);
 		}
 
 		@Example
@@ -152,11 +145,13 @@ class ArbitraryEdgeCasesTests {
 			Arbitrary<List<Integer>> arbitrary = ints.list().ofMinSize(1);
 			assertThat(values(arbitrary.edgeCases())).containsExactlyInAnyOrder(
 				Collections.singletonList(-10),
+				Collections.singletonList(-9),
 				Collections.singletonList(-2),
 				Collections.singletonList(-1),
 				Collections.singletonList(0),
 				Collections.singletonList(1),
 				Collections.singletonList(2),
+				Collections.singletonList(9),
 				Collections.singletonList(10)
 			);
 		}
@@ -170,7 +165,7 @@ class ArbitraryEdgeCasesTests {
 
 		@Example
 		void listEdgeCasesWhenFixedSize() {
-			IntegerArbitrary ints = Arbitraries.integers().between(10, 100);
+			Arbitrary<Integer> ints = Arbitraries.of(10, 100);
 			Arbitrary<List<Integer>> arbitrary = ints.list().ofSize(3);
 			assertThat(values(arbitrary.edgeCases())).containsExactlyInAnyOrder(
 				asList(10, 10, 10),
@@ -204,14 +199,16 @@ class ArbitraryEdgeCasesTests {
 			assertThat(values(arbitrary.edgeCases())).containsExactlyInAnyOrder(
 				Collections.emptySet(),
 				Collections.singleton(-10),
+				Collections.singleton(-9),
 				Collections.singleton(-2),
 				Collections.singleton(-1),
 				Collections.singleton(0),
 				Collections.singleton(1),
 				Collections.singleton(2),
+				Collections.singleton(9),
 				Collections.singleton(10)
 			);
-			assertThat(values(arbitrary.edgeCases())).hasSize(8);
+			assertThat(values(arbitrary.edgeCases())).hasSize(10);
 		}
 
 		@Example
@@ -220,37 +217,34 @@ class ArbitraryEdgeCasesTests {
 			Arbitrary<Set<Integer>> arbitrary = ints.set().ofMinSize(1);
 			assertThat(values(arbitrary.edgeCases())).containsExactlyInAnyOrder(
 				Collections.singleton(-10),
+				Collections.singleton(-9),
 				Collections.singleton(-2),
 				Collections.singleton(-1),
 				Collections.singleton(0),
 				Collections.singleton(1),
 				Collections.singleton(2),
+				Collections.singleton(9),
 				Collections.singleton(10)
 			);
 		}
 
 		@Example
 		void streamEdgeCases() {
-			IntegerArbitrary ints = Arbitraries.integers().between(-10, 10);
+			Arbitrary<Integer> ints = Arbitraries.of(-10, 10);
 			Arbitrary<Stream<Integer>> arbitrary = ints.stream();
 			Set<Stream<Integer>> streams = values(arbitrary.edgeCases());
 			Set<List<Integer>> lists = streams.stream().map(stream -> stream.collect(Collectors.toList())).collect(Collectors.toSet());
 			assertThat(lists).containsExactlyInAnyOrder(
 				Collections.emptyList(),
 				Collections.singletonList(-10),
-				Collections.singletonList(-2),
-				Collections.singletonList(-1),
-				Collections.singletonList(0),
-				Collections.singletonList(1),
-				Collections.singletonList(2),
 				Collections.singletonList(10)
 			);
-			assertThat(values(arbitrary.edgeCases())).hasSize(8);
+			assertThat(values(arbitrary.edgeCases())).hasSize(3);
 		}
 
 		@Example
 		void iteratorEdgeCases() {
-			IntegerArbitrary ints = Arbitraries.integers().between(-10, 10);
+			Arbitrary<Integer> ints = Arbitraries.of(-10, 10);
 			Arbitrary<Iterator<Integer>> arbitrary = ints.iterator();
 			Set<Iterator<Integer>> iterators = values(arbitrary.edgeCases());
 			Set<List<Integer>> lists =
@@ -264,31 +258,21 @@ class ArbitraryEdgeCasesTests {
 			assertThat(lists).containsExactlyInAnyOrder(
 				Collections.emptyList(),
 				Collections.singletonList(-10),
-				Collections.singletonList(-2),
-				Collections.singletonList(-1),
-				Collections.singletonList(0),
-				Collections.singletonList(1),
-				Collections.singletonList(2),
 				Collections.singletonList(10)
 			);
-			assertThat(values(arbitrary.edgeCases())).hasSize(8);
+			assertThat(values(arbitrary.edgeCases())).hasSize(3);
 		}
 
 		@Example
 		void arraysAreCombinationsOfElementsUpToMaxLength() {
-			IntegerArbitrary ints = Arbitraries.integers().between(-10, 10);
+			Arbitrary<Integer> ints = Arbitraries.of(-10, 10);
 			StreamableArbitrary<Integer, Integer[]> arbitrary = ints.array(Integer[].class);
 			assertThat(values(arbitrary.edgeCases())).containsExactlyInAnyOrder(
 				new Integer[]{},
 				new Integer[]{-10},
-				new Integer[]{-2},
-				new Integer[]{-1},
-				new Integer[]{0},
-				new Integer[]{1},
-				new Integer[]{2},
 				new Integer[]{10}
 			);
-			assertThat(values(arbitrary.edgeCases())).hasSize(8);
+			assertThat(values(arbitrary.edgeCases())).hasSize(3);
 		}
 
 		@Example